Grow_memory function added.
Basically, you can grow the memory on the memory block if there are
enough space to grow. The extra space will be added at the end of the
previous memory partition if there are enough space to grow or the
previous memory block will be reallocated if current address cannot be
grow.